Move trade/order Models to their own class
